Fixing up I_GetTime, renaming to I_GetTimeTicks to make it clearer, cleaning up code reliant on that function.
So I've known for a while that if you try running Rum and Raisin for a day, it eventually freezes. I finally left it running in a debugger and caught it. Turns out that I_GetTime would give bad values after a while. I fixed it to use built-in performance counters native to your OS, instead of relying on the millisecond counter provided by the Tick API. Short story there. After about 60 million ticks, the millisecond value would start wrapping when using exclusively 32-bit values. The Linux Doom source release didn't have this problem because it used a time API that returned seconds and microseconds, and expanded the tick count from there. Doing the math, it should have had an uptime of just over 700 days. Using only milliseconds means you have 1/1000th the range of using seconds as the basis, meaning you get the freeze in about 70% of a day. Like I've been seeing.

This is one that I think should be pushed to Chocco. Or at least, just rewrite the function in a similar fashion to how I did it and return a 32-bit value. That will preserve the roughly 700-day uptime behavior, although i expect it'll be out by a few ticks.

+++ src/doom/f_wipe.h
@@ -58,6 +58,6 @@ wipe_ScreenWipe
   int		y,
   int		width,
   int		height,
-  int		ticks );
+  uint64_t		ticks );
 
 #endif
