+++ src/i_video.h
@@ -134,6 +134,6 @@ extern char *window_position;
 void I_GetWindowPosition(int *x, int *y, int w, int h);
 
 // Joystic/gamepad hysteresis
-extern unsigned int joywait;
+extern uint64_t joywait;
 
 #endif
